+++ src/index.ts
@@ -1,2 +1,2 @@
-export { CustomQueryBuilder, CustomQueryBuilderError } from './CustomQueryBuilder';
+export { CustomQueryBuilder, CustomQueryBuilderError, defineQueryBuilder } from './CustomQueryBuilder';
 export type { LeftJoinsAndSelects, JoinsAndSelects } from './CustomQueryBuilder';